+++ FROST-Server.SQLjooq/src/main/java/de/fraunhofer/iosb/ilt/frostserver/persistence/pgjooq/JooqPersistenceManager.java
@@ -109,4 +109,8 @@ public interface JooqPersistenceManager extends LiquibaseUser, PersistenceManage
     public default boolean hasUpdateReturning() {
         return true;
     }
+
+    public default boolean hasDistinctOn() {
+        return true;
+    }
 }