+++ FROST-Server.SQLjooq/src/main/java/de/fraunhofer/iosb/ilt/frostserver/persistence/pgjooq/MariadbPersistenceManager.java
@@ -886,4 +886,9 @@ public class MariadbPersistenceManager extends AbstractPersistenceManager implem
         return false;
     }
 
+    @Override
+    public boolean hasDistinctOn() {
+        return false;
+    }
+
 }