+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Development version 2.5.2
 
+**Internal changes & Bugfixes**
+* MariaDB: avoid use of distinctOn on MariaDb, as not supported.
+
 
 ## Release version 2.5.1